Update Iban index and related queries
The uniqueness index for Iban in the AddressBookRecord model has been extended to include the IbanType. Previously, the uniqueness was only enforced on the Iban. Now it is on the combination of OwnerClientId, Iban, and IbanType. This change has also been reflected in related queries and services, and the database migration file has been created for this adjustment.